+++ front/app/components/modal/user/user/user-crud.directive.js
@@ -14,7 +14,7 @@ angular.module('testingItApp')
     scope: {
       type: '@',
     },
-    controller: ['$scope', '$rootScope', 'UserService', 'TestProjectCrudService', 'Restangular', function($scope, $rootScope, UserService, TestProjectCrudService, Restangular) {
+    controller: ['$scope', '$rootScope', 'UserService', 'TestProjectCrudService', 'DialogConfirmService', 'Restangular', function($scope, $rootScope, UserService, TestProjectCrudService, DialogConfirmService, Restangular) {
 
       $scope.edit = true;
       $scope.canEdit = function() { return $scope.edit; };
@@ -53,7 +53,13 @@ angular.module('testingItApp')
 
       //Delete an user
       $scope.deleteUser = function(entity){
-        UserService.deleteUser($scope, entity._id);
+
+        $scope.config = ["Are you sure?", "Do you want delete the selected user?", "Accept", "Cancel"];
+        DialogConfirmService.openDialogModal($scope.config).then(function (isOk) {
+          if (isOk){
+              UserService.deleteUser($scope, entity._id);
+          };
+        });
       };
 
       //Init uiGrid for users
